style: main layout 스타일링 변경
