+++ apps/web/src/shared/ui/index.ts
@@ -1,3 +1,2 @@
 export { Default } from './default';
 export { PageLoading } from './page-loading';
-export { ModeToggle } from './theme-toggle';
