+++ packages/shared/src/widgets/layouts/main-layout/header/theme-toggle.tsx
@@ -12,14 +12,13 @@ import {
 
 import { Moon, Sun } from 'lucide-react';
 
-/** 샘플 다크모드 */
-export function ModeToggle() {
+export function ThemeToggle() {
   const { setTheme } = useTheme();
 
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
-        <Button className="fixed right-6 bottom-6" variant="outline" size="icon">
+        <Button className="rounded-full" variant="ghost" size="icon">
           <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
           <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
           <span className="sr-only">Toggle theme</span>
