+++ packages/shared/src/widgets/layouts/main-layout/sidebar/index.tsx
@@ -1,31 +1,15 @@
-import {
-  Sidebar as ShadcnSidebar,
-  SidebarContent,
-  SidebarFooter,
-  SidebarHeader,
-} from '@it-diots/shared/ui';
-import { OAuthUser } from '@it-diots/supabase/types';
+import { Sidebar as ShadcnSidebar, SidebarContent } from '@it-diots/shared/ui';
 
+import { LogoHeader } from './logo-header';
 import { RouteGroups } from './route-groups';
-import { SidebarUser } from './sidebar-user';
 
-interface SidebarProps {
-  onSignOut: () => void;
-  userInfo: OAuthUser;
-}
-
-export function Sidebar({ onSignOut, userInfo }: SidebarProps) {
+export function Sidebar() {
   return (
     <ShadcnSidebar collapsible="icon">
-      <SidebarHeader>
-        <SidebarUser onSignOut={onSignOut} userInfo={userInfo} />
-      </SidebarHeader>
-
+      <LogoHeader />
       <SidebarContent>
         <RouteGroups />
       </SidebarContent>
-
-      <SidebarFooter />
     </ShadcnSidebar>
   );
 }
